apps/api/requirements.txt
# apps/api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydicom==2.4.3
pillow==10.1.0
watchdog==3.0.0
numpy==1.26.2
scikit-image==0.22.0
aiofiles==23.2.1
redis==5.0.1